frontend/requirements.txt
# Flask & Web
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-login>=0.6.3
werkzeug>=3.0.0
python-dotenv>=1.0.0

# AI/ML Core
transformers>=4.36.0
datasets>=2.16.0
torch>=2.0.0
accelerate>=0.25.0
google-genai>=0.1.0

# NLP & Tokenizers
sentencepiece>=0.1.99
protobuf>=4.25.0
tokenizers>=0.15.0

# Computer Vision
tensorflow>=2.13.0
opencv-python>=4.8.0
pillow>=10.0.0

# Machine Learning
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0

# Datasets
kagglehub>=0.2.0

# Audio/Speech (optional, for future STT/TTS)
speechrecognition>=3.10.0
pyttsx3>=2.90
